Modify the hygen binary
